Fix #382 #383 arrumando problemas do codeclimate
